src/third_party/blink/web_tests/paint/invalidation/subpixel-shadow-move-expected.txt

CONSOLE MESSAGE: line 13: 49.75
CONSOLE MESSAGE: line 13: 49.5
CONSOLE MESSAGE: line 13: 49.25
CONSOLE MESSAGE: line 13: 49
CONSOLE MESSAGE: line 13: 48.75
CONSOLE MESSAGE: line 13: 48.5
CONSOLE MESSAGE: line 13: 48.25
CONSOLE MESSAGE: line 13: 48
CONSOLE MESSAGE: line 13: 47.75
CONSOLE MESSAGE: line 13: 47.5
CONSOLE MESSAGE: line 13: 47.25
CONSOLE MESSAGE: line 13: 47
CONSOLE MESSAGE: line 13: 46.75
CONSOLE MESSAGE: line 13: 46.5
CONSOLE MESSAGE: line 13: 46.25
CONSOLE MESSAGE: line 13: 46
CONSOLE MESSAGE: line 13: 45.75
CONSOLE MESSAGE: line 13: 45.5
CONSOLE MESSAGE: line 13: 45.25
CONSOLE MESSAGE: line 13: 45
CONSOLE MESSAGE: line 13: 44.75
CONSOLE MESSAGE: line 13: 44.5
CONSOLE MESSAGE: line 13: 44.25
CONSOLE MESSAGE: line 13: 44
CONSOLE MESSAGE: line 13: 43.75
CONSOLE MESSAGE: line 13: 43.5
CONSOLE MESSAGE: line 13: 43.25
CONSOLE MESSAGE: line 13: 43
CONSOLE MESSAGE: line 13: 42.75
CONSOLE MESSAGE: line 13: 42.5
CONSOLE MESSAGE: line 13: 42.25
CONSOLE MESSAGE: line 13: 42
CONSOLE MESSAGE: line 13: 41.75
CONSOLE MESSAGE: line 13: 41.5
CONSOLE MESSAGE: line 13: 41.25
CONSOLE MESSAGE: line 13: 41
CONSOLE MESSAGE: line 13: 40.75
CONSOLE MESSAGE: line 13: 40.5
CONSOLE MESSAGE: line 13: 40.25
CONSOLE MESSAGE: line 13: 40
